more stuff from alpha....

+++ MSShared/encrypt.cpp
@@ -62,10 +62,10 @@ void CEncryptData1::Encrypt()
 	}
 
 	//Store checksum
-	Encrypted[m_DataSize + 0] = (byte)(CheckSum / pow(256, 3));
-	Encrypted[m_DataSize + 1] = (byte)(CheckSum / pow(256, 2));
-	Encrypted[m_DataSize + 2] = (byte)(CheckSum / pow(256, 1));
-	Encrypted[m_DataSize + 3] = (byte)((int)CheckSum % (int)pow(256, 1));
+	Encrypted[m_DataSize + 0] = (byte)(CheckSum / static_cast<int>(pow(256.0f, 3)));
+	Encrypted[m_DataSize + 1] = (byte)(CheckSum / static_cast<int>(pow(256.0f, 2)));
+	Encrypted[m_DataSize + 2] = (byte)(CheckSum / static_cast<int>(pow(256.0f, 1)));
+	Encrypted[m_DataSize + 3] = (byte)((int)CheckSum % (int)pow(256.0f, 1));
 	m_DataSize += MS_SIZE_LONG; //Set the new size of the data
 }
 
@@ -79,9 +79,9 @@ bool CEncryptData1::Decrypt()
 
 	//Retrieve checksum
 	FileCheckSum += m_pData[CheckSumPos + 3];
-	FileCheckSum += (unsigned long)(m_pData[CheckSumPos + 2] * pow(256, 1));
-	FileCheckSum += (unsigned long)(m_pData[CheckSumPos + 1] * pow(256, 2));
-	FileCheckSum += (unsigned long)(m_pData[CheckSumPos + 0] * pow(256, 3));
+	FileCheckSum += (unsigned long)(m_pData[CheckSumPos + 2] * static_cast<int>(pow(256.0f, 1)));
+	FileCheckSum += (unsigned long)(m_pData[CheckSumPos + 1] * static_cast<int>(pow(256.0f, 2)));
+	FileCheckSum += (unsigned long)(m_pData[CheckSumPos + 0] * static_cast<int>(pow(256.0f, 3)));
 	m_DataSize -= MS_SIZE_LONG; //Set the size of the data buffer
 
 	//Allocate a buffer for the original data
